Move to OpenSSL 3.1
OpenSSL 1.1.x is officially EOL in a few weeks.

Only minor change to the build process is that `LICENSE` has been
moved to `LICENSE.txt` on the 3-branch.

No changes should be required in the main Vespa code base, as
it has long since been upgraded to use 1.1.x APIs, and 3.1 is
API-compatible. Have manually tested that it builds cleanly.